+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 isotope_payu Changelog
 ======================
 
+Version 1.0.1 stable (2012-01-26)
+---------------------------------
+
+### Fixed
+- Fixed the payment method logic
+
+### Improved
+- Extended the list of available status options
+
+
 Version 1.0.0 stable (2012-01-20)
 ---------------------------------